requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==4.2.7
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-tinymce==3.6.1
django-tinymce4-lite==1.8.0
django-trumbowyg==1.1.2
gunicorn==21.2.0
jsmin==3.0.1
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
